# src/main/resources/openapi.yaml
...
openapi: "3.0.4"
...
info:
  title: Beveage Service
  version: "1.0"
  description: | # | for multi line description
    Summer is coming and we are all looking forward to having a barbecue with friends.
    For a nice barbecue, we also need some refreshing drinks. Therefore, we want to start a
    new business where we sell beverages and deliver them to people.
  contact:
    name: Group 9
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:9999/v1
    description: Development Server
# resources (one of the REST principles)
...
components:
  schemas:
    orderDTO:
      type: object
      required:
        - orderId
      properties:
        orderId:
          type: number
        positions:
          type: array
          items:
            $ref: "#/components/schemas/orderItemDTO"
        price:
          type: number
        OrderStatus:
          type: string
          enum: [SUBMITTED, PROCESSED]
      xml:
        name: order
    orderItemDTO:
      type: object
      required:
        - number
        - beverage
        - quantity
      properties:
        number:
          type: number
        beverage:
          $ref: "#/components/schemas/beverageDTO"
        quantity:
          type: number
      xml:
        name: orderItem
    beverageDTO:
      type: object
      required:
        - Bottle
      properties:
        id:
          type: number
        price:
          type: number
        inStock:
          type: boolean
        Bottle:
          $ref: "#/components/schemas/beverageDTO"
        noOfBottles:
          type: number
      xml:
        name: beverage
...
    errorMessage:
      type: object
      required:
        - errorType
      properties:
        errorType:
          description: error category, should be extended
          enum:
            - INVALID_PARAMETER
        message:
          type: string
          description: A detailed error description
...
    CustomerDTO:
      type: object
      required:
        - firstName
        - lastName
        - userName
      properties:
        userName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
paths:
...
  /employee/register:
    post:
      tags:
        - Register
      summary: create a new employee.
      parameters:
        - name: userName
          in: query
          required: true
          schema:
            type: string
        - name: firstName
          in: query
          required: true
          schema:
            type: string
        - name: lastName
          required: true
          in: query
          schema:
            type: string
      responses:
        200:
          description: Employee successfully registered.
        400:
          description: There is some internal issue while employee registration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /employee/{empId}:
    get:
      tags:
        - Register
      summary: Get employee by Id
      parameters:
        - name: empId
          required: true
          in: path
          schema:
            type: string
      responses:
        200:
          description: Retrieval of employee by specified id successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDTO"
        400:
          description: Can not retrieve employee by specified Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /employee/{employeeId}:
    description: delete a employee from the list
    parameters:
      - name: employeeId
        required: true
        in: path
        schema:
          type: number
    delete:
      tags:
        - Register
      summary: delete employee by id
      responses:
        200:
          description: delete employee by Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDTO"
        400:
          description: Employee with specific Id is deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /employee/login:
    get:
      tags:
        - login
      summary: Authenticate Employee
      parameters:
        - name: firstName
          in: query
          schema:
            type: string
            default: admin
        - name: lastName
          in: query
          schema:
            type: string
            default: admin
      responses:
        200:
          description: Authenticated Employee
        400:
          description: login failed for employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
...
  /order/{orderId}:
    description: delete an order
    parameters:
      - name: orderId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
    get:
      tags:
        - shared
      summary: Get a single order by id
      # possible responses and their parameters
      responses:
        200:
          description: get a specific order description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderDTO"
        400:
          description: A bad request initiated by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    put:
      tags:
        - employee
      summary: set order to processed
      responses:
        200:
          description: Updating an order
          content:
            application/json:
              # response schema
              schema:
                $ref: "#/components/schemas/orderDTO"
        400:
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        404:
          description: Not found
    delete:
      tags:
        - employee
      summary: Delete order
      responses:
        200:
          description: order deleted
        404:
          description: order not found
...
  /orderitem/{orderId}/{orderItemNumber}:
    # this parameter definition is valid for all subsequent methods
    parameters:
      - name: orderId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
      - name: orderItemNumber
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
      - name: newOrderAmount
        # path parameters must be always mandatory!!
        required: false
        in: query
        schema:
          type: number
    get:
      tags:
        - shared
      summary: Get a single order item of an order
      # possible responses and their parameters
      responses:
        200:
          description: get a specific orderitem description
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderItemDTO"
        400:
          description: A bad request initiated by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
    put:
      tags:
        - customer
      summary: Update specific order item
      responses:
        200:
          description: Updating an order item
          content:
            application/json:
              # response schema
              schema:
                $ref: "#/components/schemas/orderItemDTO"
        400:
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
        404:
          description: Not found
    delete:
      tags:
        - customer
      summary: Delete order Item
      responses:
        200:
          description: order item deleted
        404:
          description: order item not found
...
  /orderitem/bottle/{orderId}/{bottleId}:
    parameters:
      - name: orderId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
      - name: bottleId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
      - name: amount
        # path parameters must be always mandatory!!
        required: false
        in: query
        schema:
          type: number
    post:
      tags:
        - customer
      summary: Create an order item for a bottle
      responses:
        201:
          description: create order Item
        400:
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /orderitem/crate/{orderId}/{crateId}:
    parameters:
      - name: orderId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
      - name: crateId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
      - name: amount
        # path parameters must be always mandatory!!
        required: false
        in: query
        schema:
          type: number
    post:
      tags:
        - customer
      summary: Create an order item for a crate
      responses:
        201:
          description: create order Item
        400:
          description: Message body was empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"
  /orderitem/{orderId}:
    # this parameter definition is valid for all subsequent methods
    parameters:
      - name: orderId
        # path parameters must be always mandatory!!
        required: true
        in: path
        schema:
          type: number
    get:
      tags:
        - shared
      summary: Get a all order items of an order
      # possible responses and their parameters
      responses:
        200:
          description: get all order items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/orderItemDTO"
        400:
          description: A bad request initiated by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorMessage"